apps/worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helix-ai"
version = "0.2.0"
description = "Helix AI — Python AI layer (LLM routing, orchestration, swarms)"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiohttp>=3.10.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "mcp>=1.0.0",
    "sse-starlette>=2.0.0",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
composio = [
    "composio-core>=0.5.0",
]
bedrock = [
    "boto3>=1.35.0",
]
all = [
    "helix-ai[composio,bedrock]",
]

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
